Refactoring: to each its own
